+++ src/main/java/com/strangeone101/holoitemsapi/tracking/TrackingListener.java
@@ -2,14 +2,20 @@ package com.strangeone101.holoitemsapi.tracking;
 
 import com.strangeone101.holoitemsapi.item.BlockAbility;
 import com.strangeone101.holoitemsapi.item.CustomItemManager;
+import io.papermc.paper.event.packet.PlayerChunkLoadEvent;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.BlockBurnEvent;
+import org.bukkit.event.block.BlockDispenseEvent;
 import org.bukkit.event.block.BlockExplodeEvent;
 import org.bukkit.event.block.BlockFadeEvent;
+import org.bukkit.event.block.BlockPistonExtendEvent;
+import org.bukkit.event.block.BlockPistonRetractEvent;
 import org.bukkit.event.block.BlockPlaceEvent;
+import org.bukkit.event.inventory.InventoryCloseEvent;
+import org.bukkit.inventory.BlockInventoryHolder;
 import xyz.holocons.mc.holoitemsrevamp.HoloItemsRevamp;
 
 public class TrackingListener implements Listener {
@@ -20,7 +26,7 @@ public class TrackingListener implements Listener {
         this.plugin = plugin;
     }
 
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onBlockPlace(final BlockPlaceEvent event) {
         final var itemStack = event.getItemInHand();
 
@@ -28,11 +34,43 @@ public class TrackingListener implements Listener {
             return;
 
         plugin.getTrackingManager().track(event.getBlockPlaced(), ability);
+        ability.onBlockPlace(event, event.getBlock().getState());
     }
 
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onBlockBreak(final BlockBreakEvent event) {
-        plugin.getTrackingManager().untrack(event.getBlock());
+
+        var ability = plugin.getTrackingManager().untrack(event.getBlock());
+        if (ability == null)
+            return;
+        ability.onBlockBreak(event, event.getBlock().getState());
+    }
+
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
+    public void onBlockDispense(BlockDispenseEvent event) {
+        if (!plugin.getTrackingManager().isTracked(event.getBlock())) {
+            return;
+        }
+
+        plugin.getTrackingManager().getBlockAbility(event.getBlock())
+            .onBlockDispense(event, event.getBlock().getState());
+    }
+
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onInventoryClose(InventoryCloseEvent event) {
+        if (!(event.getInventory().getHolder(true) instanceof BlockInventoryHolder blockInventoryHolder)
+            || !plugin.getTrackingManager().isTracked(blockInventoryHolder.getBlock())) {
+            return;
+        }
+
+        plugin.getTrackingManager().getBlockAbility(blockInventoryHolder.getBlock())
+            .onInventoryClose(event, blockInventoryHolder.getBlock().getState());
+    }
+
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onPlayerChunkLoad(PlayerChunkLoadEvent event) {
+        plugin.getTrackingManager().getTrackedBlocks(event.getWorld().getUID(), event.getChunk().getChunkKey())
+            .forEach(entry -> entry.getValue().onPlayerChunkLoad(event, entry.getKey().blockState()));
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
@@ -50,5 +88,22 @@ public class TrackingListener implements Listener {
         plugin.getTrackingManager().untrack(event.getBlock());
     }
 
-    //TODO handle piston movement (maybe outright disable them when used against custom blocks?)
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
+    public void onPistonExtend(BlockPistonExtendEvent event) {
+        var hasCustomBlock = event.getBlocks().stream().anyMatch(block -> plugin.getTrackingManager().isTracked(block));
+
+        if (hasCustomBlock)
+            event.setCancelled(true);
+    }
+
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
+    public void onPistonRetract(BlockPistonRetractEvent event) {
+        if (!event.isSticky())
+            return;
+
+        var hasCustomBlock = event.getBlocks().stream().anyMatch(block -> plugin.getTrackingManager().isTracked(block));
+
+        if (hasCustomBlock)
+            event.setCancelled(true);
+    }
 }
