+++ src/main/java/xyz/holocons/mc/holoitemsrevamp/HoloItemsRevamp.java
@@ -4,7 +4,6 @@ import com.strangeone101.holoitemsapi.Keys;
 import com.strangeone101.holoitemsapi.enchantment.AnvilListener;
 import com.strangeone101.holoitemsapi.enchantment.EnchantManager;
 import com.strangeone101.holoitemsapi.enchantment.EnchantmentListener;
-import com.strangeone101.holoitemsapi.item.BlockListener;
 import com.strangeone101.holoitemsapi.recipe.CraftListener;
 import com.strangeone101.holoitemsapi.tracking.TrackingListener;
 import com.strangeone101.holoitemsapi.tracking.TrackingManager;
@@ -40,7 +39,6 @@ public final class HoloItemsRevamp extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new EnchantmentListener(), this);
         getServer().getPluginManager().registerEvents(new AnvilListener(this), this);
         getServer().getPluginManager().registerEvents(new CraftListener(this), this);
-        getServer().getPluginManager().registerEvents(new BlockListener(this), this);
         getServer().getPluginManager().registerEvents(new TrackingListener(this), this);
 
         getCommand("holoitems").setExecutor(new MainCommand(this));
